Infrastructure-message-queue: Refactor Message Broker
