+++ src/TicketR.MessageBroker.RabbitMQ/Messages/Interfaces/IRabbitMQMessageBroker.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using TicketR.MessageBroker.Infrastructure.Messages.Interfaces;
+
+namespace TicketR.MessageBroker.RabbitMQ.Messages.Interfaces
+{
+    public interface IRabbitMQMessageBroker : IMessageBroker
+    {
+    }
+}
