+++ src/TicketR.MessageBroker.RabbitMQ/Messages/Interfaces/IRabbitMQMessageHandler.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using TicketR.MessageBroker.Integrations.Interfaces;
+using TicketR.MessageBroker.RabbitMQ.Messages.Models;
+
+namespace TicketR.MessageBroker.RabbitMQ.Messages.Interfaces
+{
+    public interface IRabbitMQMessageHandler<in TMessage> : IMessageHandler<TMessage> where TMessage : RabbitMQMessage
+    {
+    }
+}
