+++ src/TicketR.MessageBroker.RabbitMQ/Messages/Models/RabbitMQMessage.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using TicketR.MessageBroker.Infrastructure.Messages.Models;
+
+namespace TicketR.MessageBroker.RabbitMQ.Messages.Models
+{
+    public class RabbitMQMessage : Message
+    {
+    }
+}
