+++ src/TicketR.MessageBroker.RabbitMQ/TicketR.MessageBroker.RabbitMQ.csproj
@@ -5,12 +5,13 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.2.0" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.1" />
     <PackageReference Include="RabbitMQ.Client" Version="5.1.0" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\TicketR.MessageBroker\TicketR.MessageBroker.csproj" />
+    <ProjectReference Include="..\TicketR.MessageBroker\TicketR.MessageBroker.Infrastructure.csproj" />
   </ItemGroup>
 
 </Project>
