+++ src/TicketR.MessageBroker/Messages/Interfaces/IMessageHandler.cs
@@ -2,11 +2,11 @@
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
-using TicketR.MessageBroker.Messages.Models;
+using TicketR.MessageBroker.Infrastructure.Messages.Models;
 
 namespace TicketR.MessageBroker.Integrations.Interfaces
 {
-    public interface IIntegrationMessageHandler<in TMessage> where TMessage : Message
+    public interface IMessageHandler<in TMessage> where TMessage : Message
     {
         Task Handle(TMessage message);
     }
